환경변수 local 추가 Kakao 로그인 API 설계 추가

+++ src/main/java/com/monkey/aggregate/temporary/service/TemporaryPostService.java
@@ -1,13 +1,14 @@
 package com.monkey.aggregate.temporary.service;
 
 import com.monkey.aggregate.temporary.repository.TemporaryPostRepository;
-import com.monkey.aggregate.temporary.view.TemporaryPostRes;
-import com.monkey.aggregate.temporary.view.TemporaryPostSaveReq;
-import com.monkey.aggregate.temporary.view.TemporaryPostUpdateReq;
+import com.monkey.aggregate.temporary.dto.TemporaryPostResponseDto;
+import com.monkey.aggregate.temporary.dto.TemporaryPostSaveDto;
+import com.monkey.aggregate.temporary.dto.TemporaryPostUpdateDto;
 import com.monkey.aggregate.temporary.domain.TemporaryPost;
 import com.monkey.aggregate.temporary.domain.TemporaryPostId;
 import com.monkey.aggregate.user.domain.UserId;
-import com.monkey.exception.ErrorCode;
+import com.monkey.aop.permission.service.PermissionService;
+import com.monkey.enums.MonkeyErrorCode;
 import com.monkey.exception.MonkeyException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,30 +19,32 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 public class TemporaryPostService {
     private final TemporaryPostRepository temporaryPostRepository;
+    private final PermissionService permissionService;
 
-    public TemporaryPostRes getPost(final TemporaryPostId postId, UserId userId) {
-        TemporaryPost post = temporaryPostRepository.findByPostIdAndUserId(postId, userId)
-                .orElseThrow(() -> new MonkeyException(ErrorCode.E100));
-
-        return new TemporaryPostRes(post.getPostId().getId(), post.getContent(), post.getCreatedAt(), post.getModifiedAt());
+    public TemporaryPostResponseDto getPost(final TemporaryPostId postId, UserId userId) {
+        TemporaryPost temporaryPost = temporaryPostRepository.findById(postId)
+                .orElseThrow(() -> new MonkeyException(MonkeyErrorCode.E100));
+        permissionService.checkPermission(userId, temporaryPost);
+        return new TemporaryPostResponseDto(temporaryPost.getPostId().getId(), temporaryPost.getContent(), temporaryPost.getCreatedAt(), temporaryPost.getModifiedAt());
     }
 
     /**
-     * @param req 임시 게시글 저장 객체
+     * @param dto 임시 게시글 저장 객체
      * @return 엔티티 인덱스 값
      */
-    public String savePost(TemporaryPostSaveReq req) {
-        TemporaryPost temporaryPost = TemporaryPost.create(req.getUserId(), req.getContent());
+    public String savePost(TemporaryPostSaveDto dto) {
+        TemporaryPost temporaryPost = TemporaryPost.builder().userId(dto.getUserId()).content(dto.getContent()).build();
         return temporaryPostRepository.save(temporaryPost).getPostId().getId();
     }
 
     /**
-     * @param req 임시 게시글 업데이트 요청객체(유저아이디, 엔티티 인덱스 값)
+     * @param dto 임시 게시글 업데이트 요청객체(유저아이디, 엔티티 인덱스 값)
      */
-    public void updatePost(TemporaryPostUpdateReq req) {
-        TemporaryPost temporaryPost = temporaryPostRepository.findById(new TemporaryPostId(req.getPostId()))
-                .orElseThrow(() -> new MonkeyException(ErrorCode.E100));
-        temporaryPost.update(req.getContent());
+    public void updatePost(TemporaryPostUpdateDto dto) {
+        TemporaryPost temporaryPost = temporaryPostRepository.findById(new TemporaryPostId(dto.getPostId()))
+                .orElseThrow(() -> new MonkeyException(MonkeyErrorCode.E100));
+        permissionService.checkPermission(dto.getUserId(), temporaryPost);
+        temporaryPost.update(dto.getContent());
     }
 
     /**
@@ -49,8 +52,9 @@ public class TemporaryPostService {
      * @param userId 유저 인덱스
      */
     public void deletePost(TemporaryPostId postId, UserId userId) {
-        TemporaryPost temporaryPost = temporaryPostRepository.findByPostIdAndUserId(postId, userId)
-                .orElseThrow(() -> new MonkeyException(ErrorCode.E100));
+        TemporaryPost temporaryPost = temporaryPostRepository.findById(postId)
+                .orElseThrow(() -> new MonkeyException(MonkeyErrorCode.E100));
+        permissionService.checkPermission(userId, temporaryPost);
         temporaryPostRepository.delete(temporaryPost);
     }
 }
